+++ lib/http/response.dart
@@ -16,28 +16,28 @@ class Response {
     this.statusCode = 200,
     this.body,
     Map<String, String>? headers,
-  }) : headers = headers ?? {'Content-Type': 'text/plain'};
+  }) : headers = headers ?? {'content-type': 'text/plain'};
 
   /// Creates a JSON response with the provided [data] and status code.
   Response.json(dynamic data, {int status = 200})
       : statusCode = status,
         body = jsonEncode(data),
-        headers = {'Content-Type': 'application/json'};
+        headers = {'content-type': 'application/json'};
 
   /// Creates an HTML response with the provided [html] string and status code.
   Response.html(String html, {int status = 200})
       : statusCode = status,
         body = html,
-        headers = {'Content-Type': 'text/html'};
+        headers = {'content-type': 'text/html'};
 
   Response.ok(this.body)
       : statusCode = 200,
-        headers = {'Content-Type': 'text/plain'};
+        headers = {'content-type': 'text/plain'};
 
   Response.redirect(String url, {int status = 302})
       : statusCode = status,
         body = null,
-        headers = {'Location': url};
+        headers = {'location': url};
 
   /// Creates a copy of this response with additional/replaced headers.
   Response withHeaders(Map<String, String> additionalHeaders) {